backend/pyproject.toml
[project]
name = "cybersim-backend"
version = "1.0.0"
description = "CyberSim educational cybersecurity simulator backend"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "httpx>=0.28.0",
    "redis[asyncio]>=5.2.0",
    "qrcode[pil]>=8.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.12",
    "email-validator>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
